Chore: add complementary prod_test using cli-testing library (#985)
* run prod_test with cli-testing

* no config

* done change prod_test

* change wait to waitForText

* cli testing as additional test

* adding catch to stop waiting changed file

* add wait instead of waitForText

---------
