+++ package.json
@@ -39,12 +39,13 @@
     "start:prod": "npm run prepack && ./bin/run",
     "test": "npm run clean && cross-env NODE_ENV=test nyc --reporter=lcov --extension .ts mocha --config .mocharc.json --forbid-only \"{src,test}/**/*.test.ts\" --exclude test/cli.test.ts",
     "cli-test": "npx bats test/bats/cli.bats",
-    "prod_test": "npx bats prod_test/test.bats",
+    "prod_test": "npm run prod_test:bats && npm run prod_test:cli",
+    "prod_test:bats": "npx bats prod_test/test.bats",
+    "prod_test:cli": "mocha prod_test/test.ts --timeout=60000",
     "version": "oclif readme && git add README.md",
     "format": "prettier --write .",
     "check-format": "prettier --check .",
-    "pkg": "npm run prepack && pkg . --compress brotli",
-    "test:e2e": "npm run prepack && mocha test/cli.test.ts --timeout=20000"
+    "pkg": "npm run prepack && pkg . --compress brotli"
   },
   "dependencies": {
     "@alicloud/dingtalk": "^1.3.53",
